refactor: move mock data for events to a columns folder
